app/requirements.txt
Django==2.2
django-webpack-loader==0.6
markdown==3.1
django-filter==2.1
django-private-storage==2.2.2
